Adding pkg: PackageJson to the TaskContext and updating all tasks to use it
This is a part of https://github.com/checkupjs/checkup/issues/400.
Also removing getPackageJson from core, so it isnt made available to the plugins, as they shouldnt need to use it.

The one task that is still implemented questionably is `ember-in-repo-addons-engines-task` - it reads all package.jsons in a repo, I think it should likely still just operate from root, and traverse and find all package.jsons, but I will revisit that implementation when I add in support for the globs