Adding checks around whether sloc supports a specific exension
The lines-of-code task now dynamically finds all types of files in your app, buckets them according to those file types, and then runs sloc for each type. If sloc does not support an exension (it exports the list of extension it supports) the LOC count will be correct, but all insights about the LOC (comments, todos, etc) will not be, so I remove those incorrect results.