Creating FilePathsArray class that extends array and offers filtering functionality (#467)
Many of the tasks were taking the paths string[] from TaskContext and using micromatch to filter by globs. Now, the paths object in TaskContext is of type FilePaths which extends array and adds function filterByGlob, to make it easier for task authors to filter the file paths array. This obscures the implementation detail of using micromatch to authors.