(bug fix) Restricting file types for LOC class, and excluding node_modules dotfiles
Previously, we were taking all "extensions" and using them to call sloc. Extensions were defined as a string after a period. However, when running this on a real app, there was a ton of random garbage getting thrown in there. I decided to just use the exensions supported by sloc, and just add
two additional (json and md). This fits our current use case, and we can re-evaluate (also re-evaluate using this subpar library) if things change and it no longer fits our use case.

Additionally, there was a bug where '**/node_modules/**' glob being passed to exclude was not catching stuff inside 'node_modules/.bin'. It seems that minimatch (which is what walksync uses under the hood) does not include dotfiles inside a glob by default, so explicitly called that out so it gets removed