Render the task category in stdout
In a previous PR, I added the concept of a string `category` to checkup, and sorted the results by category. In this PR, I actually render the category to stdout for each group as a header (bold, with the `===` before and after the string, and changed the task headers to not render the `===`, making them more sort of subheaders.