Prioritizing cli args for turning tasks on
Currently, if a task `foo` is set to "off" in config, but checkup is run via `checkup --task foo` the task doesnt run. The command line args override config for excludepaths, and it should work the same way for tasks. Now, if a task is passed in via command line arg, it will always run, regardless of config value