Adding functionality for pretty formatter to write to a file
With this refactor, the ConsoleWriter can operate in two modes. 1) write to console directly and 2) append content that wouldve been written to console to a string.
This way, the pretty formatter can write its contents to a txt file, instead of just printing to console.

This change makes it so that no two cli options are mutually exclusive! All format options work with the ouput-file option