[bugfix] Add cases to getPaths to prevent uncaught exception
This is definitely an edge case, but currently if you have a directory passed in that doesn't exist and walkSync attempts to expand it, a terminal error will be thrown. Added more explicit casing to getPaths to account for this, and manually verified that in the repo where i was running checkup (and this error was being thrown) that it now works properly.