[enhancement] Adding non-fatal errors to tasks
Rationale: Previously, when a single file could not be parsed correctly, the entire eslint-disable task would fail. We want to notify our consumers when a file can not be parsed, or when any error occurs when analyzing a single file, but we dont necessarily want that to cause the whole task to fail. It is now up to the author of a task whether they just want to throw an error when a failure is caught, or just add a nonFatalError, which are all collected at the end and pushed to the `toolExecutionNotifications`.

Implementation: Added new public function `addNonFatalError` to base-task, and collected the nonFatalErrors in task-list with the other (fatal) errors. This approach pushes the responsibility of handling failures in the analyzers onto the task authors, giving them full control of how they want to handle what happens when a file cant be analyzed (for some tasks, I imagine failure to read/parse a file to be fatal, and for the desired behavior to be a task erroring out completely. With this change, made required changes to eslint-disable task (the only task using the AstAnalyzer right now).